Remove old shader code, move particle system to entities folder
