+++ src/entities/particle.cpp
@@ -1,5 +1,5 @@
 
-#include <particle.h>
+#include <entities\particle.h>
 
 namespace rpg {
