+++ src/entities/particle_simulation.cpp
@@ -1,9 +1,9 @@
 
 #include <sys_time.h>
-#include <particle_simulation.h>
+#include <entities\particle_simulation.h>
 #include <random_manager.h>
-#include <math/base.h>
 
+#include <math/base.h>
 #include <math/vector3d.h>
 #include <math/units.h>
